+++ source/MRViewer/MRViewer.vcxproj.filters
@@ -442,6 +442,9 @@
     <ClCompile Include="MRUISaveChangesPopup.cpp">
       <Filter>Helpers</Filter>
     </ClCompile>
+    <ClCompile Include="SceneReorder.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ImGuiHelpers.h">
@@ -865,6 +868,9 @@
     </ClInclude>
     <ClInclude Include="MRUISaveChangesPopup.h">
       <Filter>Helpers</Filter>
+    </ClInclude>
+    <ClInclude Include="SceneReorder.h">
+      <Filter>Source Files</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
